Building a valid lib finally working

+++ make.py
@@ -4,31 +4,39 @@
 
 import argparse, multiprocessing, os, shutil, sys
 
-def generate(target, __folder__, cpus, builddir, bindir, compile_flags):
+def generate(target, __folder__, args, cpus, builddir, libdir, compile_flags):
 
-    if os.system("cd tensorflow" +
-    " && git clean -fdx" +
-    " && make -f tensorflow/lite/experimental/micro/tools/make/Makefile test" +
-    " && make -f tensorflow/lite/experimental/micro/tools/make/Makefile TAGS=\"openmvcam\" generate_projects"):
-        sys.exit("Make Failed...")
+    print("==============================\n Building Target - " + target + "\n==============================")
+
+    project_folder = "tensorflow/tensorflow/lite/experimental/micro/tools/make/gen/linux_x86_64/prj/softmax_test/make"
+
+    if (not os.path.isdir(project_folder)) or (not args.skip_generation):
+        if os.system("cd tensorflow" +
+        " && git clean -fdx" +
+        " && make -f tensorflow/lite/experimental/micro/tools/make/Makefile test" +
+        " && make -f tensorflow/lite/experimental/micro/tools/make/Makefile TAGS=\"openmvcam\" generate_projects"):
+            sys.exit("Make Failed...")
 
     if os.path.exists(os.path.join(builddir, target)):
         shutil.rmtree(os.path.join(builddir, target), ignore_errors = True)
-    shutil.copytree("tensorflow/tensorflow/lite/experimental/micro/tools/make/gen/linux_x86_64/prj/micro_speech/make",
+
+    shutil.copytree(project_folder,
                     os.path.join(builddir, target))
 
-    data = None
+    shutil.copytree("libm",
+                    os.path.join(builddir, target, "libm"))
+
     with open(os.path.join(builddir, target, "Makefile"), 'r') as original:
         data = original.read()
-        data = data.replace("  tensorflow/lite/experimental/micro/examples/micro_speech/main.cc", " libtf.cc")
-        data = data.replace("--std=c++11 -g -DTF_LITE_STATIC_MEMORY", "--std=c++11 -DTF_LITE_STATIC_MEMORY")
-        data = data.replace("-DNDEBUG -g -DTF_LITE_STATIC_MEMORY", "-DNDEBUG -DTF_LITE_STATIC_MEMORY")
+        data = data.replace("tensorflow/lite/experimental/micro/kernels/softmax_test.cc", "libtf.cc libm/exp.c libm/floor.c libm/frexp.c libm/round.c libm/scalbn.c")
+        data = data.replace("--O3 -DNDEBUG --std=c++11 -g -DTF_LITE_STATIC_MEMORY -DTF_LITE_DISABLE_X86_NEON", "")
+        data = data.replace("-DNDEBUG -g -DTF_LITE_STATIC_MEMORY -DTF_LITE_DISABLE_X86_NEON", "")
         data = data.replace("$(CXX) $(CXXFLAGS) -o $@ $(OBJS) $(LDFLAGS)", "arm-none-eabi-ar rcs libtf.a $(OBJS)")
 
     with open(os.path.join(builddir, target, "Makefile"), 'w') as modified:
         modified.write("CC = arm-none-eabi-gcc\n")
         modified.write("CXX = arm-none-eabi-g++\n")
-        modified.write("CCFLAGS += " + compile_flags + "\n")
+        modified.write("CCFLAGS += " + compile_flags.replace("-fno-rtti ", "").replace("-std=c++11 ", "").replace("-std=gnu++11 ", "").replace("-fpermissive ", "") + "\n")
         modified.write("CXXFLAGS += " + compile_flags + "\n")
         modified.write(data)
 
@@ -42,24 +50,67 @@ def generate(target, __folder__, cpus, builddir, bindir, compile_flags):
         " && arm-none-eabi-ar rcs libtf-mobilenet.a libtf-mobilenet.o"):
         sys.exit("Make Failed...")
 
-    if not os.path.exists((os.path.join(bindir, target))):
-        os.mkdir(os.path.join(bindir, target))
-    shutil.copy(os.path.join(builddir, target, "libtf.a"), os.path.join(bindir, target))
-    shutil.copy(os.path.join(builddir, target, "libtf-mobilenet.a"), os.path.join(bindir, target))
-    shutil.copy(os.path.join(__folder__, "libtf.h"), os.path.join(bindir, target))
-    shutil.copy(os.path.join(__folder__, "libtf-mobilenet.h"), os.path.join(bindir, target))
+    if not os.path.exists((os.path.join(libdir, target))):
+        os.mkdir(os.path.join(libdir, target))
 
-    with open(os.path.join(bindir, target, "README.txt"), 'w') as file:
+    shutil.copy(os.path.join(builddir, target, "libtf.a"), os.path.join(libdir, target))
+    shutil.copy(os.path.join(builddir, target, "libtf-mobilenet.a"), os.path.join(libdir, target))
+    shutil.copy(os.path.join(__folder__, "libtf.h"), os.path.join(libdir, target))
+    shutil.copy(os.path.join(__folder__, "libtf-mobilenet.h"), os.path.join(libdir, target))
+
+    with open(os.path.join(libdir, target, "README.txt"), 'w') as file:
         file.write("Compiled with " + compile_flags + "\n")
         file.write("Make sure to link this library with arm-none-eabi-g++ as it was written in C++.\n")
         file.write("Finally, this library outputs debugging information via printf() (so printf() must be implemented on your system).\n")
 
-def build_target(target, __folder__, cpus, builddir, bindir):
+def build_target(target, __folder__, args, cpus, builddir, libdir):
+
+    compile_flags = "-DNDEBUG " \
+                    "-DTF_LITE_DISABLE_X86_NEON " \
+                    "-DGEMMLOWP_ALLOW_SLOW_SCALAR_FALLBACK " \
+                    "-DTF_LITE_STATIC_MEMORY " \
+                    "-DTF_LITE_MCU_DEBUG_LOG " \
+                    "-D __FPU_PRESENT=1 " \
+                    "-fno-rtti " \
+                    "-fmessage-length=0 " \
+                    "-fno-exceptions " \
+                    "-fno-unwind-tables " \
+                    "-fno-builtin " \
+                    "-ffunction-sections " \
+                    "-fdata-sections " \
+                    "-funsigned-char " \
+                    "-MMD " \
+                    "-mthumb " \
+                    "-mlittle-endian " \
+                    "-mno-unaligned-access " \
+                    "-mfloat-abi=hard " \
+                    "-std=c++11 " \
+                    "-std=gnu++11 " \
+                    "-Wvla " \
+                    "-Wall " \
+                    "-Wextra " \
+                    "-Wno-unused-parameter " \
+                    "-Wno-missing-field-initializers " \
+                    "-Wno-write-strings " \
+                    "-Wno-sign-compare " \
+                    "-fno-delete-null-pointer-checks " \
+                    "-fomit-frame-pointer " \
+                    "-fpermissive " \
+                    "-nostdlib " \
+                    "-O3 "
 
     if (target == "OPENMV1") or (target == "OPENMV2") or (target == "cortex-m4"):
-        generate(target, __folder__, cpus, builddir, bindir, "-mthumb -nostartfiles -fdata-sections -ffunction-sections -mcpu=cortex-m4 -mtune=cortex-m4 -mfpu=fpv4-sp-d16 -mfloat-abi=hard")
+        generate(target, __folder__, args, cpus, builddir, libdir, compile_flags +
+                                                                   "-DARM_MATH_CM4 " +
+                                                                   "-mcpu=cortex-m4 " +
+                                                                   "-mtune=cortex-m4 " +
+                                                                   "-mfpu=fpv4-sp-d16")
     elif (target == "OPENMV3") or (target == "OPENMV4") or (target == "cortex-m7"):
-        generate(target, __folder__, cpus, builddir, bindir, "-mthumb -nostartfiles -fdata-sections -ffunction-sections -mcpu=cortex-m7 -mtune=cortex-m7 -mfpu=fpv5-sp-d16 -mfloat-abi=hard")
+        generate(target, __folder__, args, cpus, builddir, libdir, compile_flags +
+                                                                   "-DARM_MATH_CM7 " +
+                                                                   "-mcpu=cortex-m7 " +
+                                                                   "-mtune=cortex-m7 " +
+                                                                   "-mfpu=fpv5-sp-d16")
     else:
         sys.exit("Unknown target!")
 
@@ -69,6 +120,9 @@ def make():
 
     parser = argparse.ArgumentParser(description =
     "Make Script")
+
+    parser.add_argument("--skip_generation", "-s", action="store_true", default=False,
+    help="Skip TensorFlow library generation.")
 
     args = parser.parse_args()
 
@@ -81,17 +135,21 @@ def make():
     if not os.path.exists(builddir):
         os.mkdir(builddir)
 
-    bindir = os.path.join(__folder__, "bin")
+    libdir = os.path.join(__folder__, "lib")
 
-    if not os.path.exists(bindir):
-        os.mkdir(bindir)
+    if not os.path.exists(libdir):
+        os.mkdir(libdir)
 
-    build_target("OPENMV1", __folder__, cpus, builddir, bindir)
-    build_target("OPENMV2", __folder__, cpus, builddir, bindir)
-    build_target("cortex-m4", __folder__, cpus, builddir, bindir)
-    build_target("OPENMV3", __folder__, cpus, builddir, bindir)
-    build_target("OPENMV4", __folder__, cpus, builddir, bindir) 
-    build_target("cortex-m7", __folder__, cpus, builddir, bindir)
+    ###########################################################################
+
+    build_target("OPENMV1",   __folder__, args, cpus, builddir, libdir)
+    build_target("OPENMV2",   __folder__, args, cpus, builddir, libdir)
+    build_target("cortex-m4", __folder__, args, cpus, builddir, libdir)
+    build_target("OPENMV3",   __folder__, args, cpus, builddir, libdir)
+    build_target("OPENMV4",   __folder__, args, cpus, builddir, libdir) 
+    build_target("cortex-m7", __folder__, args, cpus, builddir, libdir)
+
+    print("==============================\n Done\n==============================")
 
 if __name__ == "__main__":
     make()
